Move javaagent arg to gradle file
