+++ testing-overhead-jmh/src/jmh/java/io/opentelemetry/javaagent/benchmark/servlet/app/HelloWorldController.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.benchmark.app;
+package io.opentelemetry.javaagent.benchmark.servlet.app;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
